tidy refactor
