+++ frontend/src/scenes/web-analytics/tiles/WebAnalyticsTile.tsx
@@ -72,86 +72,6 @@ export const toUtcOffsetFormat = (value: number): string => {
     return `UTC${sign}${integerPart}${formattedMinutes}`
 }
 
-// Utility function to map SQL/internal column names to UI-friendly display names
-export const getDisplayColumnName = (column: string, breakdownBy?: WebStatsBreakdown): string => {
-    // Strip the "context.columns." prefix if present
-    const baseColumn = column.replace(/^context\.columns\./, '')
-
-    // Handle known metric columns first (these should always show their metric names)
-    const metricMappings: Record<string, string> = {
-        visitors: 'Visitors',
-        views: 'Views',
-        sessions: 'Sessions',
-        bounce_rate: 'Bounce Rate',
-        session_duration: 'Session Duration',
-        total_pageviews: 'Total Pageviews',
-        unique_visitors: 'Unique Visitors',
-        scroll_gt80_percentage: 'Scroll Depth >80%',
-        rage_clicks: 'Rage Clicks',
-    }
-
-    if (metricMappings[baseColumn]) {
-        return metricMappings[baseColumn]
-    }
-
-    // Handle breakdown column - only if this is the breakdown_value column and breakdownBy is defined
-    if (baseColumn === 'breakdown_value' && breakdownBy !== undefined) {
-        switch (breakdownBy) {
-            case WebStatsBreakdown.Page:
-                return 'Path'
-            case WebStatsBreakdown.InitialPage:
-                return 'Initial Path'
-            case WebStatsBreakdown.ExitPage:
-                return 'End Path'
-            case WebStatsBreakdown.PreviousPage:
-                return 'Previous Page'
-            case WebStatsBreakdown.ExitClick:
-                return 'Exit Click'
-            case WebStatsBreakdown.ScreenName:
-                return 'Screen Name'
-            case WebStatsBreakdown.InitialChannelType:
-                return 'Channel Type'
-            case WebStatsBreakdown.InitialReferringDomain:
-                return 'Referring Domain'
-            case WebStatsBreakdown.InitialUTMSource:
-                return 'UTM Source'
-            case WebStatsBreakdown.InitialUTMCampaign:
-                return 'UTM Campaign'
-            case WebStatsBreakdown.InitialUTMMedium:
-                return 'UTM Medium'
-            case WebStatsBreakdown.InitialUTMTerm:
-                return 'UTM Term'
-            case WebStatsBreakdown.InitialUTMContent:
-                return 'UTM Content'
-            case WebStatsBreakdown.Browser:
-                return 'Browser'
-            case WebStatsBreakdown.OS:
-                return 'OS'
-            case WebStatsBreakdown.Viewport:
-                return 'Viewport'
-            case WebStatsBreakdown.DeviceType:
-                return 'Device Type'
-            case WebStatsBreakdown.Country:
-                return 'Country'
-            case WebStatsBreakdown.Region:
-                return 'Region'
-            case WebStatsBreakdown.City:
-                return 'City'
-            case WebStatsBreakdown.Timezone:
-                return 'Timezone'
-            case WebStatsBreakdown.Language:
-                return 'Language'
-            case WebStatsBreakdown.FrustrationMetrics:
-                return 'URL'
-            case WebStatsBreakdown.InitialUTMSourceMediumCampaign:
-                return 'Source / Medium / Campaign'
-        }
-    }
-
-    // Return base column name if no mapping found
-    return baseColumn
-}
-
 type VariationCellProps = { isPercentage?: boolean; reverseColors?: boolean }
 const VariationCell = (
     { isPercentage, reverseColors }: VariationCellProps = { isPercentage: false, reverseColors: false }
